requirements.txt
cryptography==2.5
Flask==1.0.2
PyJWT==1.7.1
python-dotenv==0.10.1
requests==2.21.0
